refactor(config): move pure config types into aura-config
Invert the aura/aura-config dependency so all serializable TOML
config types are owned by aura-config and re-exported from aura.

- Move LlmConfig, AgentSettings, orchestration/worker config, and
  scratchpad config out of the aura crate into aura-config, split
  into new orchestration.rs and scratchpad.rs modules.
- Relocate lenient_int into aura-config, where the moved types use it.
- Move RigBuilder the other way: from aura-config into a new
  aura::rig_builder module, since it composes runtime agents.
- Rename the old aura AgentConfig to AgentRuntimeConfig, the runtime
  build context pairing pure config with non-serializable fields
  (tool wrappers, persistence handles, session id); the TOML [agent]
  table is now the distinct aura_config::AgentConfig.
- Update all call sites and tests to the new module paths.

Ref: GH-174
