+++ crates/aura-config/Cargo.toml
@@ -4,32 +4,16 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
-# Internal crates
-aura = { path = "../aura" }
-
-# Rig dependencies
-rig-core = { workspace = true }
-
-rmcp = { workspace = true }
-futures = "0.3"
-sse-stream = "0.2"
-reqwest = { version = "0.12", features = ["stream", "json"] }
-
 # Serialization
 serde = { workspace = true }
 serde_json = { workspace = true }
 toml = { workspace = true }
 
-# Configuration
+# Configuration / CLI
 clap = { workspace = true }
-config = { workspace = true }
-
-# Async
-tokio = { workspace = true }
 
 # Error handling
 thiserror = { workspace = true }
-anyhow = { workspace = true }
 
 # Logging
 tracing = { workspace = true }
@@ -40,4 +24,3 @@ regex = { workspace = true }
 
 [dev-dependencies]
 tempfile = "3"
-
