+++ crates/aura-config/src/lenient_int.rs
