+++ crates/aura/src/config.rs
@@ -1,44 +1,36 @@
-/// Configuration structs for building Rig agents
-/// These are pure Rust structs without any TOML-specific dependencies
-use crate::orchestration::OrchestrationConfig;
-use crate::scratchpad::{ScratchpadConfig, ScratchpadToolEntry, ScratchpadToolsConfig};
-use crate::tool_wrapper::{ToolCallContext, ToolWrapper};
-use serde::{Deserialize, Serialize};
-use std::fmt;
+//! Runtime agent configuration.
+//!
+//! Pure, serializable config types live in the `aura-config` crate. This module
+//! holds the runtime build-context struct (`AgentRuntimeConfig`) that composes
+//! those TOML-parsed values with non-serializable runtime fields: tool wrappers,
+//! persistence handles, shared chat history, scratchpad runtime state, and the
+//! session id.
 
-use crate::lenient_int;
-use std::collections::HashMap;
+use crate::scratchpad::ScratchpadToolsConfig;
+use crate::tool_wrapper::{ToolCallContext, ToolWrapper};
 use std::sync::Arc;
+
+// Re-export the pure config types from `aura-config` so existing
+// `crate::config::*` paths keep resolving after the structs moved out. These
+// are the single source of truth — `aura` no longer defines its own copies.
+pub use aura_config::{
+    AgentSettings, EmbeddingConfig, LlmConfig, McpConfig, McpServerConfig, OrchestrationConfig,
+    ReasoningEffort, TodoToolsConfig, ToolsConfig, VectorStoreConfig, VectorStoreType, glob_match,
+};
 
 /// Type alias for tool context factory function.
 pub type ToolContextFactory = Arc<dyn Fn(&str) -> ToolCallContext + Send + Sync>;
 
-/// Reasoning effort level for GPT-5 models
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "lowercase")]
-pub enum ReasoningEffort {
-    Minimal,
-    Low,
-    Medium,
-    High,
-}
-
-impl fmt::Display for ReasoningEffort {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let s = match self {
-            ReasoningEffort::Minimal => "minimal",
-            ReasoningEffort::Low => "low",
-            ReasoningEffort::Medium => "medium",
-            ReasoningEffort::High => "high",
-        };
-        write!(f, "{s}")
-    }
-}
-
-/// Complete configuration for building an agent
-#[derive(Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct AgentConfig {
+/// Runtime build context for constructing agents.
+///
+/// This composes the pure TOML-parsed config types from `aura-config` with
+/// non-serializable runtime extension points: tool wrappers, persistence
+/// handles, shared chat history, scratchpad runtime state, and the session id.
+///
+/// Note: this is the renamed successor of the old `AgentConfig`. The TOML
+/// `[agent]` table shape is now `aura_config::AgentConfig` — a distinct type.
+#[derive(Default)]
+pub struct AgentRuntimeConfig {
     pub llm: LlmConfig,
     pub agent: AgentSettings,
     pub vector_stores: Vec<VectorStoreConfig>,
@@ -48,75 +40,56 @@ pub struct AgentConfig {
     /// artifacts. Scratchpad: `{memory_dir}/scratchpad/` (single-agent) or
     /// `{memory_dir}/{run_id}/iteration-{n}/scratchpad/` (orchestration).
     /// `[orchestration.artifacts].memory_dir` is still honored as a legacy fallback.
-    #[serde(default)]
     pub memory_dir: Option<String>,
     /// Orchestration mode configuration (multi-agent workflows)
-    #[serde(default)]
     pub orchestration: Option<OrchestrationConfig>,
 
-    // === Extension fields (not serialized) ===
+    // === Extension fields ===
     // These allow callers to customize agent building without modifying the builder.
     // The orchestrator uses these to inject tool wrappers and override preambles.
-    /// Optional tool wrapper applied to all MCP tools (not serialized).
+    /// Optional tool wrapper applied to all MCP tools.
     /// When set, all MCP tools are wrapped with this wrapper.
-    #[serde(skip)]
     pub tool_wrapper: Option<Arc<dyn ToolWrapper + Send + Sync>>,
 
-    /// Factory for creating ToolCallContext per tool (not serialized).
+    /// Factory for creating ToolCallContext per tool.
     /// Allows callers to inject metadata (task_id, attempt) into wrapped tool calls.
-    #[serde(skip)]
     pub tool_context_factory: Option<ToolContextFactory>,
 
-    /// Override for system prompt (not serialized).
+    /// Override for system prompt.
     /// When set, this replaces agent.system_prompt entirely.
-    #[serde(skip)]
     pub preamble_override: Option<String>,
 
     /// Glob patterns for filtering which MCP tools to include.
     /// When set, only tools matching at least one pattern are added.
     /// Supports glob syntax: `*` (any chars), `?` (single char).
     /// Empty or None means all tools are included.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub mcp_filter: Option<Vec<String>>,
 
-    /// Shared persistence for injecting `read_artifact` tool into workers (not serialized).
+    /// Shared persistence for injecting `read_artifact` tool into workers.
     /// When set, workers get access to result artifacts via the read_artifact tool.
-    #[serde(skip)]
     pub orchestration_persistence:
         Option<Arc<tokio::sync::Mutex<crate::orchestration::ExecutionPersistence>>>,
 
-    /// Shared conversation history for injecting `get_conversation_context` tool into workers (not serialized).
+    /// Shared conversation history for injecting `get_conversation_context` tool into workers.
     /// When set, workers can retrieve conversation history on demand.
-    #[serde(skip)]
     pub orchestration_chat_history: Option<Arc<Vec<rig::completion::Message>>>,
 
-    /// Session ID for grouping orchestration runs under a shared namespace (not serialized).
+    /// Session ID for grouping orchestration runs under a shared namespace.
     /// When set, persistence paths become `{memory_dir}/{session_id}/{run_id}/...`.
     /// Threaded from the web server's `chat_session_id`.
-    #[serde(skip)]
     pub session_id: Option<String>,
 
-    /// Scratchpad storage/budget handed to the 8 exploration tools (not serialized).
+    /// Scratchpad storage/budget handed to the 8 exploration tools.
     /// `Some` when scratchpad is wired up for this agent or worker.
-    #[serde(skip)]
     pub scratchpad_tools_config: Option<ScratchpadToolsConfig>,
 
-    /// Shared decision state for worker `submit_result` tool (not serialized).
+    /// Shared decision state for worker `submit_result` tool.
     /// When set, workers get the `submit_result` tool for structured output.
-    #[serde(skip)]
     pub orchestration_submit_result: Option<crate::orchestration::SubmitResultDecision>,
 }
 
-/// Configuration for TodoWrite/ReadTodos tool injection.
-#[derive(Debug, Clone, Default)]
-pub struct TodoToolsConfig {
-    /// Optional directory for persisting plans.
-    /// If None, plans are stored in-memory only.
-    pub plan_dir: Option<String>,
-}
-
 // Manual Clone implementation because Arc<dyn Trait> fields require special handling
-impl Clone for AgentConfig {
+impl Clone for AgentRuntimeConfig {
     fn clone(&self) -> Self {
         Self {
             llm: self.llm.clone(),
@@ -141,9 +114,9 @@ impl Clone for AgentConfig {
 }
 
 // Manual Debug implementation because Arc<dyn Trait> fields don't implement Debug
-impl std::fmt::Debug for AgentConfig {
+impl std::fmt::Debug for AgentRuntimeConfig {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("AgentConfig")
+        f.debug_struct("AgentRuntimeConfig")
             .field("llm", &self.llm)
             .field("agent", &self.agent)
             .field("vector_stores", &self.vector_stores)
@@ -186,466 +159,7 @@ impl std::fmt::Debug for AgentConfig {
     }
 }
 
-/// LLM provider configuration with strong typing per provider
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "provider", rename_all = "lowercase")]
-#[serde(deny_unknown_fields)]
-pub enum LlmConfig {
-    OpenAI {
-        api_key: String,
-        model: String,
-        #[serde(default)]
-        base_url: Option<String>,
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        max_tokens: Option<u64>,
-        /// Context window size in tokens.
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        context_window: Option<u64>,
-        #[serde(default)]
-        reasoning_effort: Option<ReasoningEffort>,
-        /// Controls the randomness and creativity of the llm
-        #[serde(default)]
-        temperature: Option<f64>,
-        /// Additional provider-specific parameters merged into the API request.
-        /// Provider-agnostic: works for Anthropic thinking, Gemini thinking budget, etc.
-        /// Example: `{ thinking = { type = "adaptive", budget_tokens = 8000 } }`
-        #[serde(default)]
-        additional_params: Option<serde_json::Value>,
-    },
-    Anthropic {
-        api_key: String,
-        model: String,
-        #[serde(default)]
-        base_url: Option<String>,
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        max_tokens: Option<u64>,
-        /// Context window size in tokens.
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        context_window: Option<u64>,
-        /// Controls the randomness and creativity of the llm
-        #[serde(default)]
-        temperature: Option<f64>,
-        /// Additional provider-specific parameters merged into the API request.
-        /// Provider-agnostic: works for Anthropic thinking, Gemini thinking budget, etc.
-        /// Example: `{ thinking = { type = "adaptive", budget_tokens = 8000 } }`
-        #[serde(default)]
-        additional_params: Option<serde_json::Value>,
-    },
-    Bedrock {
-        model: String,
-        region: String,
-        /// AWS profile name (optional, uses default credentials if not specified)
-        #[serde(default)]
-        profile: Option<String>,
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        max_tokens: Option<u64>,
-        /// Context window size in tokens.
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        context_window: Option<u64>,
-        #[serde(default)]
-        temperature: Option<f64>,
-        /// Additional provider-specific parameters merged into the API request.
-        /// Provider-agnostic: works for Anthropic thinking, Gemini thinking budget, etc.
-        /// Example: `{ thinking = { type = "adaptive", budget_tokens = 8000 } }`
-        #[serde(default)]
-        additional_params: Option<serde_json::Value>,
-    },
-    Gemini {
-        api_key: String,
-        model: String,
-        #[serde(default)]
-        base_url: Option<String>,
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        max_tokens: Option<u64>,
-        /// Context window size in tokens.
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        context_window: Option<u64>,
-        /// Controls the randomness and creativity of the llm
-        #[serde(default)]
-        temperature: Option<f64>,
-        /// Additional provider-specific parameters merged into the API request.
-        /// Provider-agnostic: works for Anthropic thinking, Gemini thinking budget, etc.
-        /// Example: `{ thinking = { type = "adaptive", budget_tokens = 8000 } }`
-        #[serde(default)]
-        additional_params: Option<serde_json::Value>,
-    },
-    Ollama {
-        model: String,
-        #[serde(default = "default_ollama_base_url")]
-        base_url: Option<String>,
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        max_tokens: Option<u64>,
-        /// Context window size in tokens.
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        context_window: Option<u64>,
-        /// Controls the randomness and creativity of the llm
-        #[serde(default)]
-        temperature: Option<f64>,
-        /// Parse tool calls from text output (Ollama-specific workaround).
-        ///
-        /// Some Ollama models (especially qwen3-coder) output tool calls as text
-        /// (JSON, XML, etc.) instead of using native tool_call structures. When
-        /// enabled, the agent intercepts streamed text, detects tool call patterns,
-        /// and executes them via MCP.
-        ///
-        /// Requires MCP servers to be configured - logs a warning otherwise.
-        ///
-        /// Flow: Config → Agent::maybe_wrap_with_fallback → FallbackToolExecutor
-        /// See: `fallback_tool_parser` module for supported formats.
-        #[serde(default)]
-        fallback_tool_parsing: bool,
-        /// Additional Ollama-specific parameters passed directly to the API.
-        /// Examples: seed, top_k, top_p, mirostat, etc.
-        /// See: https://github.com/ollama/ollama/blob/main/docs/modelfile.mdx#valid-parameters-and-values
-        #[serde(default)]
-        additional_params: Option<serde_json::Value>,
-    },
-    OpenRouter {
-        api_key: String,
-        model: String,
-        #[serde(default)]
-        base_url: Option<String>,
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        max_tokens: Option<u64>,
-        #[serde(default, deserialize_with = "lenient_int::deserialize_option_u64")]
-        context_window: Option<u64>,
-        #[serde(default)]
-        temperature: Option<f64>,
-        #[serde(default)]
-        additional_params: Option<serde_json::Value>,
-    },
-}
-
-fn default_ollama_base_url() -> Option<String> {
-    Some("http://localhost:11434".to_string())
-}
-
-impl Default for LlmConfig {
-    fn default() -> Self {
-        LlmConfig::OpenAI {
-            api_key: String::new(),
-            model: "gpt-4o".to_string(),
-            base_url: None,
-            reasoning_effort: None,
-            max_tokens: None,
-            context_window: None,
-            temperature: None,
-            additional_params: None,
-        }
-    }
-}
-
-impl LlmConfig {
-    /// Check if Ollama text-to-tool fallback is enabled.
-    ///
-    /// Returns true only for `LlmConfig::Ollama` with `fallback_tool_parsing = true`.
-    /// Other providers always return false (they use native tool calling).
-    pub fn is_fallback_tool_parsing_enabled(&self) -> bool {
-        matches!(
-            self,
-            LlmConfig::Ollama {
-                fallback_tool_parsing: true,
-                ..
-            }
-        )
-    }
-
-    /// Get max_tokens regardless of provider.
-    pub fn max_tokens(&self) -> Option<u64> {
-        match self {
-            LlmConfig::OpenAI { max_tokens, .. }
-            | LlmConfig::Anthropic { max_tokens, .. }
-            | LlmConfig::Bedrock { max_tokens, .. }
-            | LlmConfig::Gemini { max_tokens, .. }
-            | LlmConfig::Ollama { max_tokens, .. }
-            | LlmConfig::OpenRouter { max_tokens, .. } => *max_tokens,
-        }
-    }
-
-    /// Get context_window regardless of provider.
-    pub fn context_window(&self) -> Option<u64> {
-        match self {
-            LlmConfig::OpenAI { context_window, .. }
-            | LlmConfig::Anthropic { context_window, .. }
-            | LlmConfig::Bedrock { context_window, .. }
-            | LlmConfig::Gemini { context_window, .. }
-            | LlmConfig::Ollama { context_window, .. }
-            | LlmConfig::OpenRouter { context_window, .. } => *context_window,
-        }
-    }
-
-    /// Get additional_params regardless of provider.
-    pub fn additional_params(&self) -> Option<serde_json::Value> {
-        match self {
-            LlmConfig::OpenAI {
-                additional_params, ..
-            }
-            | LlmConfig::Anthropic {
-                additional_params, ..
-            }
-            | LlmConfig::Bedrock {
-                additional_params, ..
-            }
-            | LlmConfig::Gemini {
-                additional_params, ..
-            }
-            | LlmConfig::Ollama {
-                additional_params, ..
-            }
-            | LlmConfig::OpenRouter {
-                additional_params, ..
-            } => additional_params.clone(),
-        }
-    }
-
-    /// Get the model name regardless of provider.
-    pub fn model_name(&self) -> &str {
-        match self {
-            LlmConfig::OpenAI { model, .. }
-            | LlmConfig::Anthropic { model, .. }
-            | LlmConfig::Bedrock { model, .. }
-            | LlmConfig::Gemini { model, .. }
-            | LlmConfig::Ollama { model, .. }
-            | LlmConfig::OpenRouter { model, .. } => model,
-        }
-    }
-
-    /// Get provider name and model identifier as a tuple.
-    pub fn model_info(&self) -> (&str, &str) {
-        match self {
-            LlmConfig::OpenAI { model, .. } => ("openai", model),
-            LlmConfig::Anthropic { model, .. } => ("anthropic", model),
-            LlmConfig::Bedrock { model, .. } => ("bedrock", model),
-            LlmConfig::Gemini { model, .. } => ("gemini", model),
-            LlmConfig::Ollama { model, .. } => ("ollama", model),
-            LlmConfig::OpenRouter { model, .. } => ("openrouter", model),
-        }
-    }
-
-    /// Accessor for temperature regardless of provider.
-    pub fn temperature(&self) -> Option<f64> {
-        match self {
-            LlmConfig::OpenAI { temperature, .. }
-            | LlmConfig::Anthropic { temperature, .. }
-            | LlmConfig::Bedrock { temperature, .. }
-            | LlmConfig::Gemini { temperature, .. }
-            | LlmConfig::Ollama { temperature, .. }
-            | LlmConfig::OpenRouter { temperature, .. } => *temperature,
-        }
-    }
-}
-
-/// Agent behavior settings
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AgentSettings {
-    pub name: String,
-    pub system_prompt: String,
-    pub context: Vec<String>,
-    #[serde(default, deserialize_with = "lenient_int::deserialize_option_usize")]
-    pub turn_depth: Option<usize>,
-    /// Glob patterns for filtering which MCP tools to include.
-    /// When set, only tools matching at least one pattern are added.
-    /// Supports glob syntax: `*` (any chars), `?` (single char).
-    /// Empty or None means all tools are included.
-    /// Can be set via `[agent].mcp_filter` in TOML for single-agent configs.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub mcp_filter: Option<Vec<String>>,
-    /// Agent-level scratchpad configuration (applies to single-agent and to
-    /// workers that don't provide an override).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub scratchpad: Option<ScratchpadConfig>,
-    /// Whether this agent (single-agent or orchestration coordinator) may
-    /// invoke client-side tools advertised on the request.
-    #[serde(default)]
-    pub enable_client_tools: bool,
-    /// Glob patterns selecting which client-side tools this agent can call.
-    /// `None` or empty means all client tools are available when
-    /// `enable_client_tools = true`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub client_tool_filter: Option<Vec<String>>,
-}
-
-/// Vector store configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct VectorStoreConfig {
-    pub name: String,
-    /// Optional context string to prepend to search results for better RAG integration
-    pub context_prefix: Option<String>,
-    /// Store-type-specific configuration
-    #[serde(flatten)]
-    pub store: VectorStoreType,
-}
-
-/// Type-specific vector store configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum VectorStoreType {
-    InMemory {
-        embedding_model: EmbeddingModelConfig,
-    },
-    Qdrant {
-        embedding_model: EmbeddingModelConfig,
-        url: String,
-        collection_name: String,
-    },
-    BedrockKb {
-        knowledge_base_id: String,
-        region: String,
-        #[serde(default)]
-        profile: Option<String>,
-    },
-}
-
-/// Embedding model configuration with strong typing per provider
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "provider", rename_all = "lowercase")]
-pub enum EmbeddingModelConfig {
-    OpenAI {
-        api_key: String,
-        model: String,
-        #[serde(default)]
-        base_url: Option<String>,
-    },
-    Bedrock {
-        model: String,
-        region: String,
-        /// AWS profile name (optional, uses default credentials if not specified)
-        #[serde(default)]
-        profile: Option<String>,
-    },
-}
-
-impl EmbeddingModelConfig {
-    /// Get the provider name
-    pub fn provider(&self) -> &str {
-        match self {
-            EmbeddingModelConfig::OpenAI { .. } => "openai",
-            EmbeddingModelConfig::Bedrock { .. } => "bedrock",
-        }
-    }
-
-    /// Get the model name
-    pub fn model(&self) -> &str {
-        match self {
-            EmbeddingModelConfig::OpenAI { model, .. }
-            | EmbeddingModelConfig::Bedrock { model, .. } => model,
-        }
-    }
-}
-
-/// MCP (Model Context Protocol) configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct McpConfig {
-    pub sanitize_schemas: bool,
-    pub servers: HashMap<String, McpServerConfig>,
-}
-
-/// Individual MCP server configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "transport")]
-pub enum McpServerConfig {
-    #[serde(rename = "stdio")]
-    Stdio {
-        cmd: String,
-        args: Vec<String>,
-        env: HashMap<String, String>,
-        description: Option<String>,
-        /// Per-tool scratchpad interception thresholds, keyed by tool-name glob.
-        /// Parsed from `[mcp.servers.<name>.scratchpad]`.
-        #[serde(default)]
-        scratchpad: HashMap<String, ScratchpadToolEntry>,
-    },
-    #[serde(rename = "http_streamable")]
-    HttpStreamable {
-        url: String,
-        #[serde(default)]
-        headers: HashMap<String, String>,
-        #[serde(default)]
-        description: Option<String>,
-        #[serde(default)]
-        headers_from_request: HashMap<String, String>,
-        /// Per-tool scratchpad interception thresholds, keyed by tool-name glob.
-        /// Parsed from `[mcp.servers.<name>.scratchpad]`.
-        #[serde(default)]
-        scratchpad: HashMap<String, ScratchpadToolEntry>,
-    },
-    #[serde(rename = "sse")]
-    Sse {
-        url: String,
-        #[serde(default)]
-        headers: HashMap<String, String>,
-        #[serde(default)]
-        description: Option<String>,
-        #[serde(default)]
-        headers_from_request: HashMap<String, String>,
-        /// Per-tool scratchpad interception thresholds, keyed by tool-name glob.
-        /// Parsed from `[mcp.servers.<name>.scratchpad]`.
-        #[serde(default)]
-        scratchpad: HashMap<String, ScratchpadToolEntry>,
-    },
-}
-
-impl McpServerConfig {
-    /// Get the per-tool scratchpad thresholds for this server.
-    pub fn scratchpad(&self) -> &HashMap<String, ScratchpadToolEntry> {
-        match self {
-            McpServerConfig::Stdio { scratchpad, .. } => scratchpad,
-            McpServerConfig::HttpStreamable { scratchpad, .. } => scratchpad,
-            McpServerConfig::Sse { scratchpad, .. } => scratchpad,
-        }
-    }
-}
-
-/// Tools configuration
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ToolsConfig {
-    pub filesystem: bool,
-    pub custom_tools: Vec<String>,
-}
-
-impl Default for AgentConfig {
-    fn default() -> Self {
-        Self {
-            llm: LlmConfig::OpenAI {
-                api_key: String::new(),
-                model: "gpt-4o-mini".to_string(),
-                base_url: None,
-                max_tokens: None,
-                context_window: None,
-                temperature: None,
-                reasoning_effort: None,
-                additional_params: None,
-            },
-            agent: AgentSettings {
-                name: "Assistant".to_string(),
-                system_prompt: "You are a helpful assistant.".to_string(),
-                context: vec![],
-                turn_depth: Some(5),
-                mcp_filter: None,
-                scratchpad: None,
-                enable_client_tools: false,
-                client_tool_filter: None,
-            },
-            vector_stores: Vec::new(),
-            mcp: None,
-            tools: None,
-            memory_dir: None,
-            orchestration: None,
-            // Extension fields default to None
-            tool_wrapper: None,
-            tool_context_factory: None,
-            preamble_override: None,
-            mcp_filter: None,
-            orchestration_persistence: None,
-            orchestration_chat_history: None,
-            session_id: None,
-            scratchpad_tools_config: None,
-            orchestration_submit_result: None,
-        }
-    }
-}
-
-impl AgentConfig {
+impl AgentRuntimeConfig {
     /// Check if orchestration mode is enabled.
     ///
     /// Returns true if the `[orchestration]` section exists and `enabled = true`.
@@ -716,99 +230,19 @@ impl AgentConfig {
     }
 }
 
-/// Simple glob pattern matching for tool name filtering.
-///
-/// Supports:
-/// - `*` matches any sequence of characters (including empty)
-/// - `?` matches exactly one character
-///
-/// Examples:
-/// - `mezmo_*` matches `mezmo_logs`, `mezmo_pipelines`
-/// - `*Query*` matches `ListQuery`, `QueryKnowledgeBases`
-/// - `tool_?` matches `tool_a`, `tool_b`
-pub fn glob_match(pattern: &str, text: &str) -> bool {
-    let pattern: Vec<char> = pattern.chars().collect();
-    let text: Vec<char> = text.chars().collect();
-
-    fn match_recursive(pattern: &[char], text: &[char]) -> bool {
-        match (pattern.first(), text.first()) {
-            // Both exhausted - match!
-            (None, None) => true,
-            // Pattern exhausted but text remains - no match
-            (None, Some(_)) => false,
-            // Wildcard * - try matching zero or more characters
-            (Some('*'), _) => {
-                // Try matching zero characters (skip *)
-                if match_recursive(&pattern[1..], text) {
-                    return true;
-                }
-                // Try matching one character and continue with *
-                if !text.is_empty() && match_recursive(pattern, &text[1..]) {
-                    return true;
-                }
-                false
-            }
-            // Text exhausted but pattern has non-* remaining - check for trailing *s
-            (Some(p), None) => *p == '*' && match_recursive(&pattern[1..], text),
-            // Single character wildcard ?
-            (Some('?'), Some(_)) => match_recursive(&pattern[1..], &text[1..]),
-            // Literal character match
-            (Some(p), Some(t)) => *p == *t && match_recursive(&pattern[1..], &text[1..]),
-        }
-    }
-
-    match_recursive(&pattern, &text)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
 
     #[test]
-    fn test_glob_match_exact() {
-        assert!(glob_match("hello", "hello"));
-        assert!(!glob_match("hello", "world"));
-    }
-
-    #[test]
-    fn test_glob_match_star() {
-        assert!(glob_match("mezmo_*", "mezmo_logs"));
-        assert!(glob_match("mezmo_*", "mezmo_pipelines"));
-        assert!(glob_match("mezmo_*", "mezmo_")); // empty suffix
-        assert!(!glob_match("mezmo_*", "other_logs"));
-    }
-
-    #[test]
-    fn test_glob_match_star_middle() {
-        assert!(glob_match("*Query*", "QueryKnowledgeBases"));
-        assert!(glob_match("*Query*", "ListQuery"));
-        assert!(glob_match("*Query*", "Query"));
-        assert!(!glob_match("*Query*", "ListKnowledge"));
-    }
-
-    #[test]
-    fn test_glob_match_question() {
-        assert!(glob_match("tool_?", "tool_a"));
-        assert!(glob_match("tool_?", "tool_1"));
-        assert!(!glob_match("tool_?", "tool_ab")); // too long
-        assert!(!glob_match("tool_?", "tool_")); // too short
-    }
-
-    #[test]
-    fn test_glob_match_star_only() {
-        assert!(glob_match("*", "anything"));
-        assert!(glob_match("*", ""));
-    }
-
-    #[test]
     fn test_tool_matches_filter_none() {
-        let config = AgentConfig::default();
+        let config = AgentRuntimeConfig::default();
         assert!(config.tool_matches_filter("any_tool"));
     }
 
     #[test]
     fn test_tool_matches_filter_empty() {
-        let config = AgentConfig {
+        let config = AgentRuntimeConfig {
             mcp_filter: Some(vec![]),
             ..Default::default()
         };
@@ -817,7 +251,7 @@ mod tests {
 
     #[test]
     fn test_tool_matches_filter_patterns() {
-        let config = AgentConfig {
+        let config = AgentRuntimeConfig {
             mcp_filter: Some(vec![
                 "mezmo_*".to_string(),
                 "QueryKnowledgeBases".to_string(),
@@ -833,16 +267,16 @@ mod tests {
 
     #[test]
     fn test_client_tool_matches_filter_none() {
-        let config = AgentConfig::default();
+        let config = AgentRuntimeConfig::default();
         assert!(config.client_tool_matches_filter("Read"));
     }
 
     #[test]
     fn test_client_tool_matches_filter_empty() {
-        let config = AgentConfig {
+        let config = AgentRuntimeConfig {
             agent: AgentSettings {
                 client_tool_filter: Some(vec![]),
-                ..AgentConfig::default().agent
+                ..AgentSettings::default()
             },
             ..Default::default()
         };
@@ -851,10 +285,10 @@ mod tests {
 
     #[test]
     fn test_client_tool_matches_filter_patterns() {
-        let config = AgentConfig {
+        let config = AgentRuntimeConfig {
             agent: AgentSettings {
                 client_tool_filter: Some(vec!["Read".to_string(), "Find*".to_string()]),
-                ..AgentConfig::default().agent
+                ..AgentSettings::default()
             },
             ..Default::default()
         };
