+++ docs/breaking-changes/20260410-agent-llm-toml-configuration.md
@@ -126,4 +126,4 @@ num_predict = 2048
 
 Affected Configs: `deny_unknown_fields` on `AgentConfig` and `LlmConfig`
 
-Both `aura::config::AgentConfig` and `aura_config::config::AgentConfig` now carry `#[serde(deny_unknown_fields)]`. Any errant or stale field in either section (including the removed fields listed above) will produce a hard parse error instead of being silently ignored, calling attention to the customer.
+Both `aura_config::AgentConfig` and `aura_config::LlmConfig` carry `#[serde(deny_unknown_fields)]`. Any errant or stale field in either section (including the removed fields listed above) will produce a hard parse error instead of being silently ignored, calling attention to the customer.
